refactor: solutionTab의 상태관련 로직들을 custom hook으로 분리
