+++ src/hooks/useIsLoaded.ts
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const useIsLoaded = () => {
+  const [isLoaded, setIsLoaded] = React.useState(true);
+  return { isLoaded, setIsLoaded };
+};
+
+export { useIsLoaded };
