+++ src/pages/newTab/solution/index.tsx
@@ -2,9 +2,12 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { RecoilRoot } from 'recoil';
 import { ThemeProvider } from 'styled-components';
+
 import { theme } from '@src/styles/theme';
 import GlobalStyles from '@src/styles/global';
-import { SolutionResponse } from '@src/types/solution';
+import { useAllSolution } from '@src/hooks/solution/useAllSolution';
+import { SelectedLanguage } from '@src/types/problem/problem';
+
 import SolutionTab from './SolutionTab';
 
 const languageRegex = /&language=(.*)/;
@@ -12,22 +15,13 @@ const problemIdRegex = /num=(.*?)&name/;
 const problemNameRegex = /name=(.*?)&language/;
 
 const href = window.location.href;
-const selectedLanguage = href.match(languageRegex)![1];
+const selectedLanguage = href.match(languageRegex)![1] as SelectedLanguage;
 const problemId = href.match(problemIdRegex)![1];
 const problemName = decodeURI(href.match(problemNameRegex)![1]);
 document.title = `프로솔브 - ${problemName}`;
 
 const SolutionTabLayout = () => {
-  const [isLoaded, setIsLoaded] = React.useState(true);
-  const [solutions, setSolutions] = React.useState<SolutionResponse>({});
-
-  React.useEffect(() => {
-    (async () => {
-      const allSolutions = await getAllSolutions({ selectedLanguage, problemId });
-      setSolutions(allSolutions);
-      setIsLoaded(false);
-    })();
-  }, []);
+  const { isLoaded, solutions } = useAllSolution({ selectedLanguage, problemId });
 
   return (
     <SolutionTab>
@@ -36,33 +30,6 @@ const SolutionTabLayout = () => {
       <SolutionTab.Content isLoaded={isLoaded} solutions={solutions} />
     </SolutionTab>
   );
-};
-
-interface HrefProps {
-  selectedLanguage: string;
-  problemId: string;
-}
-type GetAllSolutionFn = ({ selectedLanguage, problemId }: HrefProps) => Promise<SolutionResponse>;
-const getAllSolutions: GetAllSolutionFn = async ({ selectedLanguage, problemId }: HrefProps) => {
-  console.log(`[Pro Solve] 문제 번호:>> ${problemId} 선택한 언어:>> ${selectedLanguage}`);
-
-  const allSolutions = await new Promise<SolutionResponse>(resolve => {
-    chrome.runtime.sendMessage(
-      {
-        method: 'getAllSolutions',
-        data: {
-          problemId,
-          selectedLanguage,
-        },
-      },
-      (response: SolutionResponse) => {
-        resolve(response);
-        console.log('[Pro Solve] 풀이한 코드 List :>>', response);
-      },
-    );
-  });
-
-  return allSolutions;
 };
 
 const root = document.createElement('div');
